refactor(owner/create): split stepper into separated file
