+++ UI/src/features/owner/pages/OwnerCreateAccomPage.tsx
@@ -1,38 +1,18 @@
 import { useState } from "react";
-import { Box, Button, Typography, Paper, Stepper, Step, StepLabel, StepButton, Alert } from "@mui/material";
-import LocationOnOutlinedIcon from "@mui/icons-material/LocationOnOutlined";
-import MeetingRoomOutlinedIcon from "@mui/icons-material/MeetingRoomOutlined";
-import KingBedOutlinedIcon from "@mui/icons-material/KingBedOutlined";
-import PhotoLibraryOutlinedIcon from "@mui/icons-material/PhotoLibraryOutlined";
-import DescriptionOutlinedIcon from "@mui/icons-material/DescriptionOutlined";
-import type { SvgIconComponent } from "@mui/icons-material";
+import { Box, Button, Typography, Paper, Alert } from "@mui/material";
 
 import PreWizardPage from "../components/PreWizard/PreWizardPage";
 import StepBasicInfoBox from "../components/Wizard/Step2/StepBasicInfoBox";
 import StepAddressBox from "../components/Wizard/Step3/StepAddressBox";
 import StepFacilityBox from "../components/Wizard/Step4/StepFacilityBox";
 import StepRoomsBox from "../components/Wizard/Step5/StepRoomBox";
-import StepImageBox from "../components/Wizard/Step6/StepImageBox";
+// import StepImageBox from "../components/Wizard/Step6/StepImageBox";
 import FacilityPanel from "../components/Wizard/Step4/FacilityPanel";
 
 import { type WizardForm } from "../types/owner.types";
 import { ERentalType, EAccommodationType } from "../../accommodation/types/accommodation.types";
-
-// ─── Step config ──────────────────────────────────────────────────────────────
-
-interface StepMeta {
-	label: string;
-	subtitle: string;
-	icon: SvgIconComponent;
-}
-
-const STEP_META: StepMeta[] = [
-	{ label: "Basic Info", subtitle: "Name & description", icon: DescriptionOutlinedIcon },
-	{ label: "Location", subtitle: "Where it is", icon: LocationOnOutlinedIcon },
-	{ label: "Facilities", subtitle: "What you offer", icon: MeetingRoomOutlinedIcon },
-	{ label: "Rooms", subtitle: "Rooms & beds", icon: KingBedOutlinedIcon },
-	{ label: "Photos", subtitle: "Images & cover", icon: PhotoLibraryOutlinedIcon },
-];
+import { CreateAccommStepper } from "../components/Wizard/CreateAccommStepper";
+import { STEP_META } from "../const/StepperMetaConst";
 
 // ─── Validation ───────────────────────────────────────────────────────────────
 
@@ -193,8 +173,8 @@ const OwnerCreateAccomPage = () => {
 				return <StepFacilityBox form={form} setForm={setForm} onSelect={(id) => setFacilityExpandedId((prev) => (prev === id ? null : id))} />;
 			case 3:
 				return <StepRoomsBox form={form} setForm={setForm} />;
-			case 4:
-				return <StepImageBox form={form} setForm={setForm} />;
+			// case 4:
+			// 	return <StepImageBox form={form} setForm={setForm} />;
 			default:
 				return null;
 		}
@@ -215,122 +195,7 @@ const OwnerCreateAccomPage = () => {
 
 			<Box display="flex" gap={3} alignItems="flex-start">
 				{/* ── Left sidebar: vertical stepper ─────────────────────────── */}
-				<Paper
-					elevation={0}
-					sx={{
-						width: 220,
-						flexShrink: 0,
-						p: 2,
-						borderRadius: 3,
-						border: "1px solid",
-						borderColor: "divider",
-						position: "sticky",
-						top: 24,
-					}}
-				>
-					<Stepper
-						nonLinear
-						activeStep={step}
-						orientation="vertical"
-						sx={{
-							"& .MuiStepConnector-line": {
-								borderColor: "divider",
-								minHeight: 20,
-								transition: "border-color 0.3s ease",
-							},
-							"& .Mui-completed .MuiStepConnector-line": {
-								borderColor: "primary.main",
-							},
-							"& .MuiStepIcon-root": {
-								color: "text.disabled",
-								transition: "color 0.2s ease",
-							},
-							"& .MuiStepIcon-root.Mui-active": { color: "primary.main" },
-							"& .MuiStepIcon-root.Mui-completed": { color: "primary.main" },
-						}}
-					>
-						{STEP_META.map((meta, i) => {
-							const isActive = step === i;
-							const isCompleted = completed.has(i);
-							const isClickable = i < step || isCompleted;
-							const Icon = meta.icon;
-
-							return (
-								<Step key={meta.label} completed={isCompleted}>
-									<StepButton
-										onClick={() => goToStep(i)}
-										disableRipple={!isClickable}
-										sx={{
-											cursor: isClickable ? "pointer" : "default",
-											borderRadius: 2,
-											py: 0.75,
-											px: 1,
-											textAlign: "left",
-											transition: "background-color 0.2s ease",
-
-											...(isClickable &&
-												!isActive && {
-													"&:hover": {
-														bgcolor: "action.hover",
-														"& .MuiStepLabel-label": { color: "primary.main" },
-														"& .step-subtitle": { color: "primary.light" },
-													},
-												}),
-
-											...(isActive && {
-												bgcolor: "primary.50",
-												borderLeft: "3px solid",
-												borderColor: "primary.main",
-												pl: "calc(8px - 3px)",
-											}),
-
-											"& .MuiStepLabel-root": { alignItems: "flex-start" },
-
-											"& .MuiStepLabel-label": {
-												fontWeight: isActive ? 700 : 500,
-												color: isActive ? "primary.main" : isCompleted ? "text.primary" : "text.disabled",
-												lineHeight: 1.2,
-												transition: "color 0.2s ease",
-											},
-
-											"& .MuiStepLabel-iconContainer": {
-												pr: 1.5,
-											},
-										}}
-									>
-										<StepLabel
-											icon={
-												<Icon
-													sx={{
-														fontSize: 20,
-														color: isActive ? "primary.main" : isCompleted ? "primary.main" : "text.disabled",
-														transition: "color 0.2s ease",
-													}}
-												/>
-											}
-										>
-											{meta.label}
-											<Typography
-												component="span"
-												className="step-subtitle"
-												variant="caption"
-												display="block"
-												sx={{
-													color: isActive ? "primary.light" : "text.disabled",
-													lineHeight: 1.2,
-													fontWeight: 400,
-													transition: "color 0.2s ease",
-												}}
-											>
-												{meta.subtitle}
-											</Typography>
-										</StepLabel>
-									</StepButton>
-								</Step>
-							);
-						})}
-					</Stepper>
-				</Paper>
+				<CreateAccommStepper step={step} completed={completed} goToStep={goToStep} />
 
 				{/* ── Right: step content ─────────────────────────────────────── */}
 				<Box flex={1} minWidth={0}>
